workflows: remove the use of nolar/setup-k3d-k3s action
After recently threatening to remove this action, I am following through
on my threat because I was working on something else and this action
suddenly stopped working for around 2hrs for no reason I could find.
When you look at the action's repo and see the shell script that sets
it up it's pretty obvious it can be implemented in a few lines of shell
if you don't need the support for various versions of k3d/k3s.

Make the change and take ownership so we can debug problems if they
occur.

Signed-off-by: John Mulligan <jmulligan@redhat.com>